// src/models/sales-detail-report.model.ts
export class SALES_DETAILS_REPORT_DTO {
  id:number=0;
  user_id: number = 0;
  user?: string = '';
  sales_number: string = '';
  sales_date: string = '';
  sales_time: string = '';
  terminal_number: string = '';
  customer_id: number =0;
  customer_code?: string = '';
  customer?: string = '';
  customer_address?:string = '';
  customer_tin?:string = '';
  table_id: number =0;
  table?: string = '';
  total_amount: number = 0;
  net_amount: number = 0;
  balance_amount: number =0;
  paid_amount:number =0;
  no_of_pax: number = 0;
  remarks?: string = '';
  status: string = '';
  is_locked?: boolean = false;
  is_billed_out?: boolean = false;
  is_cancelled?: boolean = false;
  is_printed?: boolean = false;
  senior_pwd_name?: string = '';
  senior_pwd_id?: string = '';
  discounted_pax: number = 0;

  item_id: number = 0;
  item_code: string = '';
  item_description: string = '';
  item_barcode: string = '';
  item_alias: string = '';
  item_category: string = '';
  item_cost: number = 0;
  item_image:string = '';
  is_vat_inclusive?: boolean = false;
  unit_id: number = 0;
  unit: string = '';
  unit_code: string = '';
  quantity: number = 0;
  price?: number = 0;
  discount_id: number = 0;
  discount: string = '';
  discount_rate: number = 0;
  discount_amount?: number = 0;
  net_price?: number = 0;
  amount?: number = 0;
  tax_id: number = 0;
  tax:string = '';
  tax_code:string = '';
  tax_rate:number = 0;
  tax_amount?: number = 0;
  particulars?: string = '';
}